float to double updated

+++ Level-1/1.CPP-Basic-1/10.FloatOrInt.cpp
@@ -10,7 +10,7 @@ using namespace std;
 
 int main()
 {
-    float n;
+    double n;
     cin >> n;
     int val = n;
